Create CallbackResponse and Deepgram::callback_request
When using the Callback feature, the JSON response from the API is completely different from the response of a normal request. Because of this, a new response type was needed.

Since Deepgram::prerecorded_request always returns a PrerecordedResponse, Deepgram::callback_request was made to return a CallbackResponse.

Because the Callback feature can now be determined based on which method is called, it was redudant for it also to be in the Options, so it was remove from there.