Reexport error types (#113)
Error types are defined in dependencies forcing the user to pin those
dependencies to the exact version that deepgram is using if they want to
process the errors. Reexporting the symbols allows the user to directly
use the exported symbol to analyze the error and continue using
the dependency on the version they want for other uses not related to
deepgram.

It also avoid importing extra libraries if they are just
used for deepgram error processing.